Update server.py
Invalid use of .split() function. When the data part of the message consists of user input with spaces, the _parse_msg() will raise a ValueError. Maxsplit should be 2, according to the doc of split(): Return a list of the words in the string, using sep as the delimiter string. If maxsplit is given, at most maxsplit splits are done (thus, the list will have at most maxsplit+1 elements, reference [https://docs.python.org/3/library/stdtypes.html#str.split]